refactor(JapScan): extract reader extraction to dedicated module
Address MikeZeDev review on PR #1805:
- Remove IPC/Diagnostics logging from website file (not needed in prod)
- Move #ExtractPagesFromReader to JapScan.Extract.ts (clean separation)
- Simplify FetchPages to use imported ExtractPagesFromReader + DRM fallback
- Investigated canvas rendering: 0 references in DRM.js, JapScan uses
  <img> tags + performance resource timeline (no canvas needed)

The DRM file (JapScan.DRM.js) remains untouched (compiled/obfuscated).
The extraction logic is now in a clean, testable module.
